+++ src/stats.h
@@ -124,10 +124,10 @@ inline unsigned __int64 rdtsc()
 	return 0;
 }
 #else
-inline unsigned long long rdtsc()
+inline uint64_t rdtsc()
 {
 #ifdef __amd64__
-	unsigned long long tsc;
+	uint64_t tsc;
 	asm volatile ("rdtsc; shlq $32, %%rdx; orq %%rdx, %%rax" : "=a" (tsc) :: "%rdx");
 	return tsc;
 #elif defined __ppc__
@@ -142,7 +142,7 @@ inline unsigned long long rdtsc()
 #else // i386
 	if (CPU.bRDTSC)
 	{
-		unsigned long long tsc;
+		uint64_t tsc;
 		asm volatile ("\trdtsc\n" : "=A" (tsc));
 		return tsc;
 	}
@@ -167,13 +167,13 @@ public:
 	
 	void Clock()
 	{
-		long long time = rdtsc();
+		int64_t time = rdtsc();
 		Counter -= time;
 	}
 	
 	void Unclock()
 	{
-		long long time = rdtsc();
+		int64_t time = rdtsc();
 		Counter += time;
 	}
 	
@@ -187,13 +187,13 @@ public:
 		return Counter * PerfToMillisec;
 	}
 
-	long long GetRawCounter()
+	int64_t GetRawCounter()
 	{
 		return Counter;
 	}
 
 private:
-	long long Counter;
+	int64_t Counter;
 };
 
 #endif
